moved conductor related logic from AppletBundleStore to WeStore
