+++ ui/app/src/groups/elements/installable-applets.ts
@@ -83,6 +83,7 @@ export class InstallableApplets extends LitElement {
   }
 
   renderApplets(allApplets: [Array<Entity<AppEntry>>, Array<IconSrcOption>]) {
+    console.log("ALL APPLETS: ", allApplets);
     return html`
       <install-applet-bundle-dialog
         id="applet-dialog"
