+++ ui/app/src/we-app.ts
@@ -1,6 +1,6 @@
 import { provide } from "@lit-labs/context";
 import { state, customElement } from "lit/decorators.js";
-import { AdminWebsocket, AppWebsocket } from "@holochain/client";
+import { AdminWebsocket, AppWebsocket, encodeHashToBase64 } from "@holochain/client";
 import { LitElement, html, css } from "lit";
 import { invoke } from "@tauri-apps/api";
 
@@ -96,6 +96,7 @@ export class WeApp extends LitElement {
 
     const appStoreAppInfo = await appWebsocket.appInfo({"installed_app_id": "AppStore-0.6.x"});
     const devhubAppInfo = await appWebsocket.appInfo({"installed_app_id": "DevHub-0.6.x"});
+    // console.log("MY DEVHUB PUBLIC KEY: ", encodeHashToBase64(devhubAppInfo.agent_pub_key));
 
     getProvisionedCells(appStoreAppInfo).map(([roleName, cellInfo]) => console.log(`Appstore network seed: ${getCellNetworkSeed(cellInfo)}`));
     if (devhubAppInfo) getProvisionedCells(devhubAppInfo).map(([roleName, cellInfo]) => console.log(`Appstore network seed: ${getCellNetworkSeed(cellInfo)}`));
